% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcScaleDistMats}
\alias{calcScaleDistMats}
\title{Calculate matrices for NSGP covariance function}
\usage{
calcScaleDistMats(A_List, coords)
}
\arguments{
\item{A_List}{List of anisotropy matrices}

\item{coords}{Matrix of input coordinates (covariates)}
}
\value{
A list of ScaleMat and DistMat matrices
}
\description{
Calculates matrices 'ScaleMat' and 'DistMat', which are used to 
obtain NSGP covariance matrices
}
\examples{
## See examples in vignette:
# vignette("nsgpr", package = "GPFDA")
}
