% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgp.functions.R
\name{CGPprediction}
\alias{CGPprediction}
\title{Prediction of Convolved Gaussian process}
\usage{
CGPprediction(
  train = NULL,
  Data.train = NULL,
  Data.new = NULL,
  noiseFreePred = F,
  meanModel = NULL,
  mean = 0
)
}
\arguments{
\item{train}{List resulting from training which is an 'mgpr' object.}

\item{Data.train}{List of training data}

\item{Data.new}{List of test input data}

\item{noiseFreePred}{Logical. If TRUE, predictions will be noise-free.}

\item{meanModel}{Model for mean.}

\item{mean}{Is the mean taken out when analysis? Default to be 0, which
assumes the mean is zero. if assume mean is a constant, mean=1; if assume
mean is a linear trend, mean='t'.}
}
\value{
A list containing  \describe{ 
\item{pred.mean}{Mean of predictions}
\item{pred.sd}{Standard deviation of predictions}
\item{noiseFreePred}{Logical. If TRUE, predictions are noise-free.}
}
}
\description{
Prediction of Convolved Gaussian process
}
