% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgp.functions.R
\name{CGPR}
\alias{CGPR}
\title{Convolved (multivariate) Gaussian process regression}
\usage{
CGPR(Data, m = NULL, meanModel = 0, mean = NULL)
}
\arguments{
\item{Data}{List of two elements: (1) response variable; and (2) input
variables}

\item{m}{If Subset of Data is to be used, m denotes the subset size and
cannot be larger than the total sample size. Default set to NULL.}

\item{meanModel}{Model for mean.}

\item{mean}{Is the mean taken out when analysis? Default to be 0, which
assumes the mean is zero. if assume mean is a constant, mean=1; if assume
mean is a linear trend, mean='t'.}

\item{Type}{of mean.}
}
\value{
A list containing: \describe{ 
\item{fitted.mean }{Fitted value of training data } 
\item{fitted.sd }{Standard deviation of the fitted value of training data} 
\item{X}{Original input variables}
\item{Y}{Original response}
\item{idx}{Original response}

\item{Cov}{Original response}
\item{mean}{CHECK}
\item{meanModel}{CHECK  Mean function type}
\item{meanLinearModel}{CHECK Mean model}
}
}
\description{
Convolved (multivariate) Gaussian process regression
}
\references{
Shi, J. Q., and Choi, T. (2011), ``Gaussian Process Regression
 Analysis for Functional Data'', CRC Press.
}
