\name{co2}
\alias{co2}
\docType{data}
\title{
co2 data set for real example.
}
\description{
data.frame with two variables, the first one is the response, second one is the time. The original data has 612 samples, but 5 of them are missing, which is removed from our sample. 
}
\usage{allnorm}
\format{
  A data frame with 607 observations on the following 2 variables.
}

\details{
Data used in the real data example, see demo 'co2'. It is obtained from 
  http://cdiac.esd.ornl.gov/ftp/trends/co2/maunaloa.co2.

Atmospheric CO2 values (ppmv) derived from in situ air samples collected at Mauna Loa, Hawaii, USA
}