% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.functions6.R
\name{cov.linear}
\alias{cov.linear}
\title{Linear covariance function}
\usage{
cov.linear(hyper, Data, Data.new = NULL)
}
\arguments{
\item{hyper}{The hyperparameters. Must be a list with certain names.}

\item{Data}{The input data. Must be a vector or a matrix.}

\item{Data.new}{The data for prediction. Must be a vector or a matrix.
Default to NULL.}
}
\value{
Covariance matrix
}
\description{
Linear covariance function
}
