% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DistanceMatrix}
\alias{DistanceMatrix}
\alias{DistMat}
\alias{DistMat_sq}
\alias{DistMatLinear}
\alias{DistMatLinear_sq}
\title{Calculate generalised distances}
\usage{
DistMat(input, inputNew, A, power)

DistMat_sq(input, A, power)

DistMatLinear(input, inputNew, A)

DistMatLinear_sq(input, A)
}
\arguments{
\item{input}{Vector of the input coordinate t}

\item{inputNew}{Vector of the input coordinate t'}

\item{A}{Anisotropy matrix A}

\item{power}{Power value p}
}
\value{
A matrix
}
\description{
Calculates the generalised distance between vectors t and t'
using an anisotropy matrix A.

\itemize{
\item \code{DistMat} and \code{DistMat_sq} calculate:
\deqn{ [(t - t')^{p/2}]^T A  (t - t')^{p/2}  }
\item \code{DistMatLinear} and \code{DistMatLinear_sq} calculate:
\deqn{ t^T A t' }
}
}
\details{
The \code{DistMat_sq} and \code{DistMatLinear_sq} functions are 
used when input vectors t and t' are identical, returning a symmetric matrix. \cr \cr
When \code{DistMat} and \code{DistMat_sq} functions are used in 
powered exponential kernels, power=1 gives the exponential kernel and 
power=2 gives the squared exponential one. \cr \cr
\code{DistMatLinear} and \code{DistMatLinear_sq} functions are used in the 
linear covariance kernel.
}
