% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPFDA.R
\docType{package}
\name{GPFDA}
\alias{GPFDA}
\title{GPFDA: A package for Gaussian Process Regression for Functional Data Analysis.}
\description{
Gaussian Process Regression for Functional Data Analysis
}
\section{Mypackage functions}{

The GPFDA functions ...
}

