% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpfr.functions6.R
\name{gpfr}
\alias{gpfr}
\title{Gaussian process functional regression (GPFR) model}
\usage{
gpfr(
  response,
  time,
  uReg = NULL,
  fxReg = NULL,
  gpReg = NULL,
  fyList = NULL,
  uCoefList = NULL,
  fxList = NULL,
  fxCoefList = NULL,
  concurrent = TRUE,
  Cov = "pow.ex",
  gamma = 2,
  nu = 1.5,
  hyper = NULL,
  NewHyper = NULL,
  useGradient = T,
  rel.tol = 1e-10,
  trace.iter = 5,
  fitting = TRUE
)
}
\arguments{
\item{response}{Response data. It can be an 'fd' object or a matrix with
'nrep' rows and 'n' columns.}

\item{time}{Input 't' of functional objects. It is a numeric vector of
length 'n'.}

\item{uReg}{Scalar covariates for the FR model. It should be a matrix with
'nrep' rows.}

\item{fxReg}{Functional covariates for the FR model. It can be a matrix with
'nrep' rows  and 'n' columns, an 'fd' object, or a list of matrices or 'fd'
objects.}

\item{gpReg}{Covariates in the GP model. It should be a matrix, a numeric vector,
an 'fd' object, a list of matrices or a list of 'fd' objects.}

\item{fyList}{A list to control the smoothing of response.}

\item{uCoefList}{A list to control the smoothing of the regression
coefficient function of the scalar covariates in the FR model.}

\item{fxList}{A list to control the smoothing of functional covariates in the
FR model.}

\item{fxCoefList}{A list to control the smoothing of the regression
coefficient function of functional covariates in the functional concurrent
model.}

\item{concurrent}{Logical. If TRUE (default), concurrent functional
regression will be carried out; otherwise, the full functional regression
will be carried out.}

\item{Cov}{Covariance function(s) to use. Options are: 'linear', 'pow.ex',
'rat.qu', and 'matern'. Default to 'power.ex'.}

\item{gamma}{Power parameter used in powered exponential kernel function. It
must be 0<gamma<=2.}

\item{nu}{Smoothness parameter of the Matern class. It must be a positive
value.}

\item{hyper}{Vector of initial hyperparameters. Default to NULL.}

\item{NewHyper}{Vector of names of new hyperparameters from the customized
kernel function.}

\item{useGradient}{Logical. If TRUE, first derivatives will be used in the
optimization.}

\item{rel.tol}{Relative tolerance passed to nlminb(). Default to be 1e-10.}

\item{trace.iter}{Print the processing of iterations of optimization.}

\item{fitting}{Logical. If TRUE, GPFR fitting is carried out on the supplied data. Default to TRUE.}
}
\value{
A list containing: \describe{ 
  \item{fitted.mean}{Fitted values obtained by GPFR model}
  \item{fitted.sd}{Standard deviation of the fitted values obtained by GPFR model}
  \item{FRmodelsList}{List of fitted FR models used for the mean function}
  \item{residFR}{Residuals obtained in FR model which were used in GPR model.} 
  \item{CovFun}{Covariance kernel used in GPR model} \item{gamma}{Parameter 'gamma' used
  in Gaussian process with powered exponential kernel} \item{nu}{Parameter
  'nu' used in Gaussian process with Matern kernel} 
  \item{hyper}{Marginal ML estimates of hyperparameters of GPR model}
  \item{rawResponse}{Raw response data}
  \item{uTrain}{Training scalar covariates used in FR model}
  \item{fxTrain}{Training functional covariates used in FR model}
  \item{gpTrain}{Training covariates used in GPR model}
  \item{time}{Input time 't'} 
  \item{...}{Other objects used for 'print', 'summary', and 'predict' methods} 
  }
}
\description{
Use functional regression (FR) model for the mean structure and Gaussian
Process (GP) for the covariance structure. \cr \cr Let 'n' be the number of
time points 't' of functional objects and 'nrep' the number of independent
replications in the sample.
}
\details{
\code{fyList} is a list with the following items:

  \itemize{ \item \code{nbasis}: number of basis functions used in
  smoothing, default to be the minimum between 23 and 20\% of the number of time points. \item \code{norder}:
  order of the functional curves; default to be 6. \item \code{bSpline}:
  logical. If TRUE (default), B-splines basis is used; otherwise, Fourier
  basis is used. \item \code{Pen}: default to be c(0,0), meaning that the
  penalty is only applied to the second order derivative of the curve, with
  no penalty for the zero-th and first order derivatives of the curve. \item
  \code{lambda}: smoothing parameter for the penalty, default to be 1e-4. }
  \code{fxList} is similar to \code{fyList}. However, it is a list of lists
  to allow for different specifications for each functional covariate if
  there are multiple ones.

  \code{uCoefList} and \code{fxCoefList} are similar to
  each other. Each one is expected to be a list of lists. If a list of one
  element is provided, then the items of this element are applied to each of
  the functional coefficients of scalar covariates and of functional covariates,
  respectively.

  \itemize{ 
  \item \code{nbasis}: nnumber of basis functions used in smoothing, default to be
  less than or equal to 19. \item  \code{norder}: order of the functional
  curves; default to be 6. \item \code{bSpline}: logical. If TRUE (default),
  B-splines basis is used; otherwise, Fourier basis is used. \item
  \code{Pen}: default to be c(0,0). \item \code{lambda}: smoothing parameter
  for the penalty, default to be 1e4. \item \code{bivar}:logical. Used for
  non-concurrent models; if TRUE, bivariate basis will be used; if FALSE
  (default), normal basis will be used; see details in
  \code{\link[fda]{bifdPar}}. \item \code{lambdas}: smoothing parameter for
  the penalty of the additional basis, default to be 1. } Note that all items
  have default settings.
}
\examples{
## See examples in vignette:
# vignette("gpfr", package = "GPFDA")
}
\references{
\itemize{ \item Ramsay, J., and Silverman, B. W. (2006),
  ``Functional Data Analysis'', 2nd ed., Springer, New York. \item Shi, J.
  Q., and Choi, T. (2011), ``Gaussian Process Regression Analysis for
  Functional Data'', CRC Press. }
}
