% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpfr.functions6.R
\name{gpfrpred}
\alias{gpfrpred}
\title{Prediction of the Gaussian Process using functional regression}
\usage{
gpfrpred(
  object,
  TestData,
  NewTime = NULL,
  lReg = NULL,
  fReg = NULL,
  gpReg = NULL,
  GP_predict = TRUE
)
}
\arguments{
\item{object}{The result from training with class`gpfda'. If missing,
function stops running.}

\item{TestData}{Test input data. Must be matrix or fd object.}

\item{NewTime}{New time for test data. If NULL, default setting will be
applied.}

\item{lReg}{The test scale data for functional regression with scale
covariates.}

\item{fReg}{The test functional data for functional regression with
functional covariates.}

\item{gpReg}{List of three items. The names of the items must be `response',
`input', `time'. For type I prediction, `response' is the observed response
for a new batch, `input' is the observed functional covariates for a new
batch, `time' is the observation time for the previous two. If NULL, type
II prediction will be carried out.}

\item{GP_predict}{Logical. If true, GP prediction is carried out, otherwise
functional prediction is carried out. Default to TRUE.}
}
\value{
A list containing: \describe{ \item{ypred}{matrix of predicted value
  with confidence interval. First column is the fitted values, second and
  third are the confidence interval.} \item{ypred.mean}{The mean value of the
  prediction.} \item{ypred.sd}{The standard deviation of the prediction.}
  \item{time}{time of test data} \item{object}{all items trained from gpfr if
  exists} }
}
\description{
Predict the new points in Gaussian Process using the training results
}
\details{
Two types of prediction are supplied. Type one is the new batch has
  a few observations, type two is the new batch has no observations.
}
\examples{
## See examples in vignette:
# vignette("gpfr", package = "GPFDA")
}
\references{
\itemize{ \item Ramsay, J., and Silverman, B. W. (2006),
  ``Functional Data Analysis'', 2nd ed., Springer, New York. \item Shi, J.
  Q., and Choi, T. (2011), ``Gaussian Process Regression Analysis for
  Functional Data'', CRC Press. }
}
