% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgp.functions.R
\name{mgpCovMat}
\alias{mgpCovMat}
\title{Calculate a multivariate Gaussian processes covariance matrix given a
vector of hyperparameters}
\usage{
mgpCovMat(Data, hp)
}
\arguments{
\item{Data}{List of two elements: 'input' and 'response'. The element 'input'
is a list of N vectors, where each vector represents the input covariate
values for a particular output. The element 'response' is the corresponding
list of N matrices (if there are multiple realisations) or vectors (for a
single realisation) representing the response variables.}

\item{hp}{Vector of hyperparameters}
}
\value{
Covariance matrix
}
\description{
Calculate a multivariate Gaussian processes covariance matrix given a
vector of hyperparameters
}
\examples{
## See examples in vignette:
# vignette("mgpr", package = "GPFDA")
}
\references{
Shi, J. Q., and Choi, T. (2011), ``Gaussian Process Regression
 Analysis for Functional Data'', CRC Press.
}
