% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsgp.functions.R
\name{NSGPCovMat_Asym}
\alias{NSGPCovMat_Asym}
\title{Calculate an asymmetric NSGP covariance matrix}
\usage{
NSGPCovMat_Asym(
  hp,
  input,
  inputNew,
  nBasis = 5,
  corrModel = corrModel,
  gamma = NULL,
  nu = NULL,
  cyclic = NULL,
  whichTau = NULL
)
}
\arguments{
\item{hp}{Vector of hyperparameters estimated by function NSGPR.}

\item{input}{List of Q input variables (see details).}

\item{inputNew}{List of Q test set input variables.}

\item{nBasis}{Number of B-spline basis functions in each coordinate direction 
along which parameters change}

\item{corrModel}{Correlation function specification used for g(.). It can be 
either "pow.ex" or "matern".}

\item{gamma}{Power parameter used in powered exponential kernel function.
It must be 0<gamma<=2.}

\item{nu}{Smoothness parameter of the Matern class. It must be a positive value.}

\item{cyclic}{Logical vector of dimension Q to define which covariates are cyclic}

\item{whichTau}{Logical vector of dimension Q identifying which input coordinates}
}
\value{
A (asymmetric) covariance matrix
}
\description{
Calculate an asymmetric NSGP covariance matrix
}
\references{
Konzen, E., Shi, J. Q. and Wang, Z. (2020) "Modeling
  Function-Valued Processes with Nonseparable and/or Nonstationary Covariance
  Structure" (https://arxiv.org/abs/1903.09981)
}
