% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.functions6.R
\name{plot.gpr}
\alias{plot.gpr}
\title{Plot GPR model for either training or prediction}
\usage{
\method{plot}{gpr}(
  x,
  fitted = F,
  ylab = "y",
  xlab = "t",
  ylim = NULL,
  whichInput = 1,
  realisation = NULL,
  alpha = 0.05,
  colourData = "red",
  colourPred = "black",
  lwd = 0.5,
  cex.points = 2,
  cex.lab = 10,
  cex.axis = 10,
  cex.main = 15,
  main = NULL,
  ...
)
}
\arguments{
\item{x}{The 'gpr' object from either training or predicting of the Gaussian
Process.}

\item{fitted}{Logical. Plot fitted values or not. Default to FALSE. If FALSE,
plot the predictions.}

\item{ylab}{Title for the y-axis.}

\item{xlab}{Title for the x-axis.}

\item{ylim}{Range value for y-axis.}

\item{whichInput}{Which input coordinate direction should be at the x-axis. 
Default to be the first one.}

\item{realisation}{Integer identifying which realisation should be plotted
(if there are multiple).}

\item{alpha}{Significance level used for 'fitted' or 'prediction'. Default is 0.05.}

\item{colourData}{Colour for training data.}

\item{colourPred}{Colour for predictive mean for the new curve.}

\item{lwd}{Graphical parameter.}

\item{cex.points}{Graphical parameter.}

\item{cex.lab}{Graphical parameter.}

\item{cex.axis}{Graphical parameter.}

\item{cex.main}{Graphical parameter.}

\item{main}{Title of the plot.}

\item{...}{Not used here.}
}
\value{
A plot.
}
\description{
Plot Gaussian process for a given an object of class 'gpr'.
}
\examples{
## See examples in vignette:
# vignette("gpr_ex1", package = "GPFDA")
}
