% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgp.functions.R
\name{plot.mgpr}
\alias{plot.mgpr}
\title{Plot predictions of a multivariate Gaussian Process regression
model}
\usage{
\method{plot}{mgpr}(
  x,
  Data.obs,
  Data.new,
  i,
  ylim = NULL,
  mfrow = NULL,
  cex = 1,
  cex.lab = 1,
  cex.axis = 1,
  ...
)
}
\arguments{
\item{x}{An object of class 'mgpr'}

\item{Data.obs}{List of observed data}

\item{Data.new}{List of test data}

\item{i}{Index identifying which realisation should be plotted.}

\item{ylim}{Range of y-axis}

\item{mfrow}{Graphical parameter}

\item{cex}{Graphical parameter}

\item{cex.lab}{Graphical parameter}

\item{cex.axis}{Graphical parameter}

\item{...}{Graphical parameters passed to plot().}
}
\value{
A plot showing predictions of each element of the multivariate
  process.
}
\description{
Plot predictons of each element of the multivariate Gaussian Process for a
given an object of class 'mgpr'.
}
\examples{
## See examples in vignette:
# vignette("mgpr", package = "GPFDA")
}
