% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgp.functions.R
\name{plotCGPprediction}
\alias{plotCGPprediction}
\title{Plot predictions of a convolved (multivariate) Gaussian Process regression
model}
\usage{
plotCGPprediction(
  train,
  Data.train,
  Data.new,
  i,
  ylim = NULL,
  mfrow = NULL,
  cex = 1,
  cex.lab = 1,
  cex.axis = 1
)
}
\arguments{
\item{train}{An 'mgpr' object}

\item{Data.train}{List of training data}

\item{Data.new}{List of test data}

\item{i}{Index identifying which realisation should be plotted.}

\item{ylim}{Range of y-axis}

\item{mfrow}{Graphical parameter}

\item{cex}{Graphical parameter}

\item{cex.lab}{Graphical parameter}

\item{cex.axis}{Graphical parameter}
}
\value{
A plot showing predictions of each element of the multivariate
  process.
}
\description{
Plot predictons of each element of the multivariate Gaussian Process for a
given an object of class 'gpr'.
}
\examples{
## See examples in vignette:
# \code{vignette("cgpr", package = "GPFDA")}
}
