% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgp.functions.R
\name{plotmgpCovFun}
\alias{plotmgpCovFun}
\title{Plot auto- or cross-covariance function of a multivariate Gaussian process}
\usage{
plotmgpCovFun(
  type = "Cov",
  output,
  outputp,
  Data,
  hp,
  idx,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  xlim = NULL,
  cex.lab = 10,
  cex.axis = 10
)
}
\arguments{
\item{type}{Character. It can be either 'Cov' (for covariance function) or
'Cor' (for corresponding correlation function).}

\item{output}{Integer identifying one element of the multivariate process.}

\item{outputp}{Integer identifying one element of the multivariate process.
If 'output' and 'outputp' are the same, the auto-covariance function will
be plotted. Otherwise, the cross-covariance function between 'output' and
'outputp' will be plotted.}

\item{Data}{List containing data.}

\item{hp}{Vector of hyperparameters}

\item{idx}{Index vector identifying to which output the elements of
concatenated vectors correspond to.}

\item{xlab}{Graphical parameter.}

\item{ylab}{Graphical parameter.}

\item{ylim}{Graphical parameter.}

\item{xlim}{Graphical parameter.}

\item{cex.lab}{Graphical parameter.}

\item{cex.axis}{Graphical parameter.}
}
\value{
A plot.
}
\description{
Plot auto- or cross-covariance function of a multivariate Gaussian process
}
\examples{
## See examples in vignette:
# vignette("mgpr", package = "GPFDA")
}
