% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpfr.functions6.R
\name{print.gpfr}
\alias{print.gpfr}
\title{Print method for 'gpfr' objects}
\usage{
\method{print}{gpfr}(x, ...)
}
\arguments{
\item{x}{A 'gpfr' object.}

\item{...}{Not used here.}
}
\value{
Returns a summary of an object of class 'gpfr'.
}
\description{
Print method for 'gpfr' objects
}
\seealso{
\link[GPFDA]{gpfr}
}
