% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.functions6.R
\name{print.gpr}
\alias{print.gpr}
\title{Print method for 'gpr' objects}
\usage{
\method{print}{gpr}(x, ...)
}
\arguments{
\item{x}{A 'gpr' object.}

\item{...}{Not used here.}
}
\value{
Returns a summary of an object of class 'gpr'.
}
\description{
Print method for 'gpr' objects
}
\seealso{
\link[GPFDA]{gpr}
}
