% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgp.functions.R
\name{print.mgpr}
\alias{print.mgpr}
\title{Print method for 'mgpr' objects}
\usage{
\method{print}{mgpr}(x, ...)
}
\arguments{
\item{x}{An 'mgpr' object.}

\item{...}{Not used here.}
}
\value{
Returns a summary of an object of class 'mgpr'.
}
\description{
Print method for 'mgpr' objects
}
\seealso{
\link[GPFDA]{mgpr}
}
