% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpfr.functions6.R
\name{summary.gpfr}
\alias{summary.gpfr}
\title{Summary method for 'gpfr' objects}
\usage{
\method{summary}{gpfr}(object, ...)
}
\arguments{
\item{object}{A 'gpfr' object.}

\item{...}{Not used here.}
}
\value{
Some fitting results of the GPFR model.
}
\description{
Summary method for 'gpfr' objects
}
\seealso{
\link[GPFDA]{gpfr}
}
