% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.functions6.R
\name{summary.gpr}
\alias{summary.gpr}
\title{Summary method for 'gpr' objects}
\usage{
\method{summary}{gpr}(object, ...)
}
\arguments{
\item{object}{A 'gpr' object.}

\item{...}{Not used here.}
}
\value{
Some fitting results of the GPR model.
}
\description{
Summary method for 'gpr' objects
}
\seealso{
\link[GPFDA]{gpr}
}
