% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgp.functions.R
\name{summary.mgpr}
\alias{summary.mgpr}
\title{Summary method for 'mgpr' objects}
\usage{
\method{summary}{mgpr}(object, ...)
}
\arguments{
\item{object}{An 'mgpr' object.}

\item{...}{Not used here.}
}
\value{
Some fitting results of the MGPR model.
}
\description{
Summary method for 'mgpr' objects
}
\seealso{
\link[GPFDA]{mgpr}
}
