% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgp.functions.R
\name{CGPCovMat}
\alias{CGPCovMat}
\title{Calculation of a Convolved Gaussian processes covariance matrix given a
vector of hyperparameters}
\usage{
CGPCovMat(Data, hp)
}
\arguments{
\item{Data}{List of two elements: (1) response variable; and (2) input
variables}
}
\value{
Covariance matrix
}
\description{
Calculation of a Convolved Gaussian processes covariance matrix given a
vector of hyperparameters
}
\references{
Shi, J. Q., and Choi, T. (2011), ``Gaussian Process Regression
 Analysis for Functional Data'', CRC Press.
}
