% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgp.functions.R
\name{CGPprediction}
\alias{CGPprediction}
\title{Prediction of Convolved Gaussian process}
\usage{
CGPprediction(
  train = NULL,
  Data.train = NULL,
  Data.new = NULL,
  noiseFreePred = F,
  meanModel = NULL,
  mu = 0
)
}
\arguments{
\item{train}{List resulting from training which is an 'mgpr' object.}

\item{Data.train}{List of training data}

\item{Data.new}{List of test input data}

\item{noiseFreePred}{Logical. If TRUE, predictions will be noise-free.}

\item{meanModel}{Type of mean function for each output. It can be \describe{
\item{0}{Zero mean function} \item{1}{Constant mean function to be
estimated} \item{'t'}{Linear model for mean function} \item{'avg'}{The
average across replications is used as the mean function. This is only used
if there are more than two realisations observed at the same input
coordinate values.} } Default to 0. If argument 'mu' is specified, then
'meanModel' will be set to 'userDefined'.}

\item{mu}{Vector having the mean function values defined by the user. Its
length must be the same as the sample size, that is, ncol(response).}
}
\value{
A list containing  \describe{ 
\item{pred.mean}{Mean of predictions}
\item{pred.sd}{Standard deviation of predictions}
\item{noiseFreePred}{Logical. If TRUE, predictions are noise-free.}
}
}
\description{
Prediction of Convolved Gaussian process
}
