% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.functions6.R
\name{cov.matern}
\alias{cov.matern}
\title{Stationary Matern covariance function}
\usage{
cov.matern(hyper, input, input.new = NULL, nu)
}
\arguments{
\item{hyper}{The hyperparameters. Must be a list with certain names.}

\item{input}{The input data. Must be a vector or a matrix.}

\item{input.new}{The data for prediction. Must be a vector or a matrix.
Default to NULL.}

\item{nu}{Smoothness parameter of the Matern class. Must be a positive value.}
}
\value{
Covariance matrix
}
\description{
Stationary Matern covariance function
}
\details{
The names for the hyperparameters should be:"linear.a" for linear
  covariance function, "pow.ex.w", "pow.ex.v" for power exponential,
  "matern.w", "matern.v" for Matern, "rat.qu.s", "rat.qu.a" for
  rational quadratic, "vv" for white noise. All hyper parameters should be in
  one list.
}
