% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.functions6.R
\name{cov.matern}
\alias{cov.matern}
\title{Stationary Matern covariance function}
\usage{
cov.matern(hyper, Data, Data.new = NULL, nu)
}
\arguments{
\item{hyper}{The hyperparameters. Must be a list with certain names.}

\item{Data}{The input data. Must be a vector or a matrix.}

\item{Data.new}{The data for prediction. Must be a vector or a matrix.
Default to NULL.}

\item{nu}{Smoothness parameter of the Matern class. Must be a positive value.}
}
\value{
Covariance matrix
}
\description{
Stationary Matern covariance function
}
