% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.functions6.R
\name{cov.rat.qu}
\alias{cov.rat.qu}
\title{Stationary rational quadratic covariance function}
\usage{
cov.rat.qu(hyper, input, input.new = NULL)
}
\arguments{
\item{hyper}{The hyperparameters. Must be a list with certain names.}

\item{input}{The input data. It must be either a matrix, where each column
represents a covariate, or a vector if there is only one covariate.}

\item{input.new}{The test input for prediction. Must be a vector or a matrix.
Default to NULL. If NULL, input.new will be identical to 'input'.}
}
\value{
A covariance matrix
}
\description{
Stationary rational quadratic covariance function
}
\details{
The names for the hyperparameters should be:"linear.a" for linear
  covariance function, "pow.ex.w", "pow.ex.v" for power exponential,
  "matern.w", "matern.v" for Matern, "rat.qu.s", "rat.qu.a" for
  rational quadratic, "vv" for white noise. All hyperparameters should be in
  one list.
}
