% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpfr.functions6.R
\name{gpfr}
\alias{gpfr}
\title{Gaussian Process for functional data.}
\usage{
gpfr(
  response,
  lReg = NULL,
  fReg = NULL,
  fyList = NULL,
  fbetaList_l = NULL,
  fxList = NULL,
  fbetaList = NULL,
  concurrent = TRUE,
  fbetaList_f = NULL,
  gpReg = NULL,
  hyper = NULL,
  Cov = c("pow.ex", "linear"),
  gamma = 2,
  nu = 1.5,
  useGradient = T,
  time = NULL,
  NewHyper = NULL,
  accuracy = c("high", "normal", "low"),
  trace.iter = 5,
  fitting = FALSE,
  rPreIdx = FALSE
)
}
\arguments{
\item{response}{The training response. can be an fd object or a matrix with
nrow samples, ncol time points}

\item{lReg}{The input variable for functional linear regression with scale
covariates. Expected to be a matrix with nrow samples.}

\item{fReg}{The input variable for functional linear regression with
functional covariates. Expected to be a matrix with nrow samples, or an fd
object, or a list of matrices or fd objects.}

\item{fyList}{The list to control the smoothing of response. See details for
more info.}

\item{fbetaList_l}{The list to control the smoothing of beta for functional
regression with scale covariates. See details for more info.}

\item{fxList}{The list to control the smoothing of functional covariates for
functional regression with functional covariates. See details for more
info.}

\item{fbetaList}{The list to control the smoothing of functional covariates
for functional regression with functional covariates and scale response.
Not available for now.}

\item{concurrent}{Logical. If True concurrent functional regression will be
carried out, otherwise the full functional regression will be carried out.}

\item{fbetaList_f}{The list to control the smoothing of beta for functional
regression with functional covariates. See details for more info.}

\item{gpReg}{Data for training Gaussian Process. Expecting matrix, fd object,
list of matrices or list of fd objects.}

\item{hyper}{Vector of initial hyperparameters. Default to NULL.}

\item{Cov}{Kernel function or covariance function type(s).}

\item{gamma}{Power parameter that cannot be estimated by simple non-linear
optimization.}

\item{nu}{Smoothness parameter of the Matern class. Must be a positive value.}

\item{useGradient}{Logical. If TRUE, first derivatives will be used in the
optimization.}

\item{time}{Time used in globle setting for functional objects.}

\item{NewHyper}{Vector of the names of the new hyperparameters from 
customized kernel function.}

\item{accuracy}{Optimization accuracy. Default to be high.}

\item{trace.iter}{Print the processing of iterations of optimization.}

\item{fitting}{Is fitting required or not. Default to FALSE.}

\item{rPreIdx}{Logical. If TRUE, do random selected index for
pre-optimization, otherwise use fixed index.}
}
\value{
A list containing: \describe{ \item{hyper}{Estimated hyperparameters}
  \item{I}{A vector of estimated standard deviation of hyperparameters}
  \item{modellist}{List of models fitted before Gaussian process}
  \item{CovFun}{Covariance function} \item{gamma}{gamma used in Gaussian
  process powered exponential kernel} \item{init_resp}{Initial response
  value} \item{resid_resp}{Residual after the fitted values from models has
  been taken out} \item{fitted}{Fitted values} \item{fitted.sd}{Standard
  deviation of the fitted values} \item{ModelType}{The model applied in the
  function.} \item{lTrain}{Training data for functional regression with
  scalper covariates} \item{fTrain}{Training data for functional regression
  with functional covariates} \item{mfTrainfd}{List of fd objects that from
  training data for functional regression with functional covariates}
  \item{gpTrain}{Training data for Gaussian Process} \item{time}{Time used in
  training in Gaussian Process} \item{iuuL}{Inverse of covariance matrix for
  lReg} \item{iuuF}{Inverse of covariance matrix for fReg}
  \item{fittedFM}{Fitted values from functional regression}
  \item{fyList}{fyList used in the function} }
}
\description{
Use functional regression to be the mean structure and Gaussian Process to be
the covariance structure.
}
\details{
fyList is a list with items: `time': a sequence of time points
  default to be 100 points from 0 to 1; `nbasis': number of basis functions
  used in smoothing, default to be less or equal to 23; `norder': the order
  of the functional curves default to be 6, `bSpline': logical, if True,
  b-spline is used, otherwise use Fourier basis, default to be True; `Pen':
  default to be c(0,0), means that the penalty is on the second order
  derivative of the curve, since the weight for zero-th and first order
  derivatives of the curve are zero, `lambda':default to be 1e-4, the
  smoothing parameter for the penalty.

  fxList is a list of lists which are similar to fyList. Because it may
  contain different information for more than one functional covariates.

  fbetaList, fbetaList_l and fbetaList_f are similar to each other. They are
  also expected to be list of lists. The items in each sub-list are: `rtime':
  range of time, default to be 0 and 1; `nbasis': number of basis functions
  used in smoothing, default to be less or equal to 19; norder: the order of
  the functional curves default to be 6;`bSpline': logical, if True, b-spline
  is used, otherwise use Fourier basis, default to be True; `Pen': default to
  be c(0,0);`lambda':default to be 1e4; 'bivar':logical, if True, the
  bivariate basis will be calculated, otherwise normal basis, default to be
  False; `lambdas':the smoothing parameter for the penalty of the additional
  basis, default to be 1e4.

  Note that user only write the item they need to change in the list, all
  items have default settings. See example below.
}
\examples{
## See examples in vignette:
# vignette("gpfr", package = "GPFDA")
}
\references{
\itemize{ \item Ramsay, J., and Silverman, B. W. (2006),
  ``Functional Data Analysis'', 2nd ed., Springer, New York. \item Shi, J.
  Q., and Choi, T. (2011), ``Gaussian Process Regression Analysis for
  Functional Data'', CRC Press. }
}
