% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.functions6.R
\name{gppredict}
\alias{gppredict}
\title{Prediction using Gaussian Process}
\usage{
gppredict(
  train = NULL,
  Data.new = NULL,
  noiseFreePred = F,
  hyper = NULL,
  Data = NULL,
  Y = NULL,
  mSR = NULL,
  Cov = NULL,
  gamma = NULL,
  nu = NULL,
  meanModel = 0,
  mean = 0
)
}
\arguments{
\item{train}{Result from training which is a 'gpr' object. Default to be
NULL. If NULL, do training based on the other given arguments; if TRUE,
other arguments (except for Data.new) will replaced by NULL; if FALSE, only
do prediction based on the other given arguments.}

\item{Data.new}{The test data. Must be a vector or a matrix.}

\item{noiseFreePred}{Logical. If TRUE, predictions will be noise-free.}

\item{hyper}{The hyperparameters. Default to NULL. If not NULL, then must be
a list with certain names.}

\item{Data}{The input data from train data. Matrix or vectors are both
acceptable. Some data.frames are not acceptable.}

\item{Y}{--------- CHECK   Same as Data????  The input data from train data.
Matrix or vectors are both acceptable. Some data.frames are not acceptable.
--------------}

\item{mSR}{Subset size m if Subset of Regressors method is used. It must be
smaller than the total sample size.}

\item{Cov}{Covariance function(s) to use. Default to 'power.ex'.}

\item{gamma}{Power parameter used in powered exponential kernel function.}

\item{nu}{Smoothness parameter of the Matern class. Must be a positive value.}

\item{meanModel}{Type of mean.}

\item{mean}{Is the mean taken out when analysis? Default to be 0, which
assumes the mean is zero. if assume mean is a constant, mean=1; if assume
mean is a linear trend, mean='t'.}
}
\value{

}
\description{
Prediction using Gaussian Process
}
