% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpfr.functions6.R
\name{plot.gpfr}
\alias{plot.gpfr}
\title{Plot Gaussian Process regression with functional mean for either training or
prediction}
\usage{
\method{plot}{gpfr}(x, type = c("raw", "fitted", "prediction"), ...)
}
\arguments{
\item{x}{Plot Gaussian Process with functional mean for training or
predicting with 'gpfr' class object.}

\item{type}{Function provides three types of plots: raw, fitted and
prediction.}

\item{...}{Graphical parameters passed to plot().}
}
\value{
A plot
}
\description{
Plot Gaussian Process regression with functional mean for either training or
prediction
}
\examples{
## See examples in vignette:
# vignette("gpfr", package = "GPFDA")
}
