% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpfr.functions6.R
\name{plot.gpfr}
\alias{plot.gpfr}
\title{Plot Gaussian Process regression with functional mean for either training or
prediction}
\usage{
\method{plot}{gpfr}(
  x,
  type = c("raw", "meanFunction", "fitted", "prediction"),
  ylab = "y",
  xlab = "t",
  ...
)
}
\arguments{
\item{x}{Plot Gaussian Process with functional mean for training or
predicting with 'gpfr' class object.}

\item{type}{Function provides the following types of plots: raw,
meanFunction, fitted and prediction.}

\item{ylab}{Title for the y axis.}

\item{xlab}{Title for the x axis.}

\item{...}{Other graphical parameters passed to plot().}
}
\value{
A plot
}
\description{
Plot Gaussian Process regression with functional mean for either training or
prediction
}
\examples{
## See examples in vignette:
# vignette("gpfr", package = "GPFDA")
}
