% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgp.functions.R
\name{plot.mgpr}
\alias{plot.mgpr}
\title{Plot predictions of MGPR model}
\usage{
\method{plot}{mgpr}(
  x,
  DataObs,
  DataNew,
  nCol = NULL,
  xlab = "t",
  ylabs = NULL,
  ylims = NULL,
  alpha = 0.05,
  colourData = "black",
  colourPred = "red",
  lwd = 0.5,
  cex.points = 2,
  cex.lab = 10,
  cex.axis = 10,
  ...
)
}
\arguments{
\item{x}{An object of class 'mgpr'.}

\item{DataObs}{List of observed data.}

\item{DataNew}{List of test data.}

\item{nCol}{Number of columns that the grid of plots should have.}

\item{xlab}{Title for the x-axis.}

\item{ylabs}{Title for the y-axis.}

\item{ylims}{Ranges of values for y-axis for each of the N response variables. It must be a matrix with N rows and 2 columns.}

\item{alpha}{Significance level used for 'fitted' or 'prediction'. Default is 0.05.}

\item{colourData}{Colour for training data.}

\item{colourPred}{Colour for predictive mean for the new curves.}

\item{lwd}{Graphical parameter.}

\item{cex.points}{Graphical parameter.}

\item{cex.lab}{Graphical parameter.}

\item{cex.axis}{Graphical parameter.}

\item{...}{Not used here.}
}
\value{
A plot showing predictions of each element of the multivariate
  process.
}
\description{
Plot predictons of each of the N elements of the multivariate Gaussian Process 
for a given an object of class 'mgpr'.
}
\examples{
## See examples in vignette:
# vignette("mgpr", package = "GPFDA")
}
